Add each_move utility to make strategies easier
each_mode loops over all possible moves and runs the given callback
passing the game_manager and the current direction to it.  This makes it
easier to see what the board may look like once a given move is
executed.

After the callback is called, it reverts the game back to its previous
state.