Add max square sum with two ahead lookup
A variation of the max square sum in which the best direction is chosen
based on whether next next move will end up with a higher sum.  This
allows for suboptimal moves to be made in preparation for a better
future move (e.g. line up two high value blocks to merge them on the
next move).